refactor: move DeckUtils to :anki-common
For the widget feature module extraction

`libanki` is the logical package - these would have been in the
`libanki` module if it weren't for the 1:1 mapping to pylib

Part of 20737
