+++ AnkiDroid/src/main/java/com/ichi2/widget/cardanalysis/CardAnalysisWidgetConfig.kt
@@ -32,8 +32,8 @@ import com.ichi2.anki.common.utils.ext.unregisterReceiverSilently
 import com.ichi2.anki.databinding.ActivityCardAnalysisWidgetConfigBinding
 import com.ichi2.anki.dialogs.registerDeckSelectedHandler
 import com.ichi2.anki.dialogs.startDeckSelection
-import com.ichi2.anki.isCollectionEmpty
 import com.ichi2.anki.launchCatchingTask
+import com.ichi2.anki.libanki.isCollectionEmpty
 import com.ichi2.anki.model.SelectableDeck
 import com.ichi2.anki.startup.ensureStorageReady
 import com.ichi2.anki.ui.internationalization.sentenceCase
