+++ AnkiDroid/src/main/java/com/ichi2/widget/deckpicker/DeckPickerWidget.kt
@@ -30,8 +30,8 @@ import com.ichi2.anki.R
 import com.ichi2.anki.analytics.UsageAnalytics
 import com.ichi2.anki.common.coroutines.applicationScope
 import com.ichi2.anki.common.crashreporting.CrashReportService
-import com.ichi2.anki.isCollectionEmpty
 import com.ichi2.anki.libanki.DeckId
+import com.ichi2.anki.libanki.isCollectionEmpty
 import com.ichi2.anki.pages.DeckOptionsDestination
 import com.ichi2.widget.ACTION_UPDATE_WIDGET
 import com.ichi2.widget.AnalyticsWidgetProvider
